+++ lib/ruby_challenge/models/trip.rb
@@ -2,8 +2,26 @@
 
 module RubyChallenge
   class Trip
-    def initialize
-      @potato = nil
+    attr_reader :segments, :beginning, :destination, :errors
+
+    def initialize(segments:, beginning:, destination:)
+      @segments = segments
+      @beginning = beginning
+      @destination = destination
+      @errors = []
+    end
+
+    def valid?
+      @errors << "segments is not valid" if !@segments.is_a?(Array) || @segments.empty?
+      @errors << "beginning is not valid" if @beginning.nil? || !@beginning.is_a?(Time)
+      @errors << "destination is not valid" if @destination.nil? || @destination.length != 3
+      @errors.empty?
+    end
+
+    def to_s
+      result = "TRIP to #{destination}\n"
+      segments.each { |segment| result += "#{segment}\n" }
+      result
     end
   end
 end