+++ spec/spec_helper.rb
@@ -1,5 +1,6 @@
 # frozen_string_literal: true
 
+require "byebug"
 Dir.glob("./spec/support/**/*.rb").each { |f| require f }
 
 RSpec.configure do |config|